server/ml/requirements.txt
# Flask==3.1.2
# numpy==2.2.6
# opencv-python==4.12.0.88
# pandas==2.3.3
# pillow==11.3.0
# scikit-learn==1.7.2
# ultralytics==8.3.231

blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
filelock==3.19.1
Flask==3.1.2
fonttools==4.60.1
fsspec==2025.9.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==11.3.0
polars==1.35.2
polars-runtime-32==1.35.2
psutil==7.1.3
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
setuptools==70.2.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.9.1+cpu
torchvision==0.24.1+cpu
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.231
ultralytics-thop==2.0.18
urllib3==2.5.0
Werkzeug==3.1.3